Collapse allocation formwizards in to a single view
