+++ caseworker/templates/case/slices/summary.html
@@ -51,7 +51,7 @@
 				{% lcs 'cases.CasePage.DetailsTab.RAISED_BY' %}
 			</dt>
 			<dd class="app-case__summary-list__value">
-				<a id="hmrc_organisation" class="govuk-link govuk-link--no-visited-state" href="{% url 'organisations:organisation' case.data.hmrc_organisation.id %}?return_to={{ CURRENT_PATH }}&return_to_text={{ case.reference_code }}">{{ case.data.hmrc_organisation.name }}</a>
+				<a id="hmrc_organisation" class="govuk-link govuk-link--no-visited-state" href="{% url 'organisations:organisation' case.data.hmrc_organisation.id %}?return_to={{ CURRENT_PATH|urlencode }}&return_to_text={{ case.reference_code }}">{{ case.data.hmrc_organisation.name }}</a>
 			</dd>
 		</div>
 	{% endif %}
@@ -159,7 +159,7 @@
 		</dt>
 		<dd class="app-case__summary-list__value">
 			{% if case.case_officer %}
-				<a id="link-case-officer" class="govuk-link govuk-link--no-visited-state" href="{% url 'users:user' case.case_officer.id %}?return_to={{ CURRENT_PATH }}&return_to_text={{ case.reference_code }}">
+				<a id="link-case-officer" class="govuk-link govuk-link--no-visited-state" href="{% url 'users:user' case.case_officer.id %}?return_to={{ CURRENT_PATH|urlencode }}&return_to_text={{ case.reference_code }}">
 					{{ case.case_officer|username }}
 				</a>
 			{% else %}
@@ -187,7 +187,7 @@
 					{% for user in users %}
 						<li>
                             <div class="app-case__summary-list__value__fullwidth">
-							    <a href="{% url 'users:user' user.id %}?return_to={{ CURRENT_PATH }}&return_to_text={{ case.reference_code }}" class="govuk-link govuk-link--no-visited-state">{{ user|username }}</a>
+							    <a href="{% url 'users:user' user.id %}?return_to={{ CURRENT_PATH|urlencode }}&return_to_text={{ case.reference_code }}" class="govuk-link govuk-link--no-visited-state">{{ user|username }}</a>
                                 <a href="{% url 'cases:remove-case-assignment' queue.id case.id %}?assignment_id={{user.assignment_id}}" class="govuk-link govuk-link--no-visited-state app-case__summary-list__value__item-action">Remove</a>
                             </div>
 						</li>
@@ -199,7 +199,7 @@
 				</span>
 			{% endfor %}
             <div class="app-case__summary-list__value__actions">
-			    <a id="link-change-assigned-users" class="govuk-link govuk-link--no-visited-state" href="{% url 'cases:add-case-assignment' queue.id case.id %}">
+			    <a id="link-change-assigned-users" class="govuk-link govuk-link--no-visited-state" href="{% url 'queues:case_assignments_assign_user' queue.id %}?cases={{case.id}}&return_to={{CURRENT_PATH|urlencode}}">
 			    	{% lcs 'cases.CasePage.DetailsTab.ADD_ASSIGNED_USER' %}
 			    </a>
             </div>
