+++ unit_tests/caseworker/queues/test_views.py
@@ -13,6 +13,7 @@ from core import client
 from core.exceptions import ServiceError
 
 from caseworker.cases.helpers.case import LU_POST_CIRC_FINALISE_QUEUE_ALIAS, LU_PRE_CIRC_REVIEW_QUEUE_ALIAS
+from caseworker.queues.forms import CaseAssignmentUsersForm, CaseAssignmentQueueForm
 
 queue_pk = "59ef49f4-cf0c-4085-87b1-9ac6817b4ba6"
 
@@ -778,3 +779,266 @@ def test_activity_updates(url, authorized_client):
     assert "additional line3" not in updates[0]
     assert "Joe Bloggs" in updates[1]
     assert "applied for a licence." in updates[1]
+
+
+def test_case_assignments_add_user_system_queue(
+    authorized_client,
+    data_queue,
+    data_standard_case,
+    data_assignment,
+    mock_standard_case_with_assignments,
+    mock_queue,
+    mock_gov_users,
+):
+
+    case = data_standard_case
+    url_base = reverse(
+        "queues:case_assignments_assign_user",
+        kwargs={
+            "pk": data_queue["id"],
+        },
+    )
+    url = f"{url_base}?cases={case['case']['id']}"
+    response = authorized_client.get(url)
+    assert response.status_code == 200
+    context = response.context
+    assert isinstance(context["form"], CaseAssignmentUsersForm)
+    # Flow initiated from a system queue, so expect the SELECT_QUEUE step next
+    assert context["wizard"]["steps"].next == "SELECT_QUEUE"
+
+    html = BeautifulSoup(response.content, "html.parser")
+    assert "Which users do you want to assign to this case?" in html.find("h1").get_text()
+
+
+@pytest.fixture
+def mock_users_team_queues_list(requests_mock, gov_uk_user_id, data_queue):
+    url = client._build_absolute_uri(f"/users/{gov_uk_user_id}/team-queues/")
+    return requests_mock.get(url=url, json={"queues": [[data_queue["id"], "Some Queue"]]})
+
+
+def test_case_assignments_add_user_system_queue_submit_success(
+    authorized_client,
+    data_queue,
+    data_standard_case,
+    data_assignment,
+    mock_standard_case_with_assignments,
+    mock_queue,
+    mock_gov_users,
+    mock_put_assignments,
+    mock_users_team_queues_list,
+    mock_standard_case_ecju_queries,
+    mock_standard_case_assigned_queues,
+    mock_standard_case_documents,
+    mock_standard_case_additional_contacts,
+    mock_standard_case_activity_filters,
+):
+    user_id = "1f288b81-2c26-439f-ac32-2a43c8b1a5cb"
+    case = data_standard_case
+    url_base = reverse("queues:case_assignments_assign_user", kwargs={"pk": data_queue["id"]})
+    url = f"{url_base}?cases={case['case']['id']}"
+    data = {
+        "case_assignments_case_assignee-current_step": "SELECT_USERS",
+        "SELECT_USERS-users": [user_id],
+        "SELECT_USERS-note": "foobar",
+    }
+    # POST step 1
+    response = authorized_client.post(url, data)
+    assert response.status_code == 200
+    context = response.context
+    assert isinstance(context["form"], CaseAssignmentQueueForm)
+    assert context["wizard"]["steps"].current == "SELECT_QUEUE"
+
+    html = BeautifulSoup(response.content, "html.parser")
+    assert "Select a team queue to add the case to" in html.find("h1").get_text()
+
+    data = {
+        "case_assignments_case_assignee-current_step": "SELECT_QUEUE",
+        "SELECT_QUEUE-queue": data_queue["id"],
+    }
+    # POST step 2
+    response = authorized_client.post(url, data, follow=True)
+    assert response.status_code == 200
+    assert response.redirect_chain[-1][0] == f"/queues/{data_queue['id']}/"
+    messages = [str(msg) for msg in response.context["messages"]]
+    expected_message = "Case adviser was added successfully"
+    assert messages == [expected_message]
+    assert mock_put_assignments.called
+    assert mock_put_assignments.last_request.json() == {
+        "case_assignments": [
+            {
+                "case_id": data_standard_case["case"]["id"],
+                "users": [user_id],
+            }
+        ],
+        "note": "foobar",
+        "remove_existing_assignments": False,
+    }
+
+    html = BeautifulSoup(response.content, "html.parser")
+    assert expected_message in html.select("div.app-snackbar__content")[0].get_text()
+
+
+def test_case_assignments_add_user_system_queue_submit_validation_error(
+    authorized_client,
+    data_queue,
+    data_standard_case,
+    data_assignment,
+    mock_standard_case_with_assignments,
+    mock_queue,
+    mock_gov_users,
+    mock_put_assignments,
+    mock_users_team_queues_list,
+    mock_standard_case_ecju_queries,
+    mock_standard_case_assigned_queues,
+    mock_standard_case_documents,
+    mock_standard_case_additional_contacts,
+    mock_standard_case_activity_filters,
+):
+    user_id = "1f288b81-2c26-439f-ac32-2a43c8b1a5cb"
+    case = data_standard_case
+    url_base = reverse("queues:case_assignments_assign_user", kwargs={"pk": data_queue["id"]})
+    url = f"{url_base}?cases={case['case']['id']}"
+    data = {
+        "case_assignments_case_assignee-current_step": "SELECT_USERS",
+        "SELECT_USERS-note": "foobar",
+    }
+    # POST step 1 - validation error
+    response = authorized_client.post(url, data)
+    assert response.status_code == 200
+    assert response.context["form"]["users"].errors == ["Select a user to allocate"]
+
+    data = {
+        "case_assignments_case_assignee-current_step": "SELECT_USERS",
+        "SELECT_USERS-users": [user_id],
+        "SELECT_USERS-note": "foobar",
+    }
+    # POST step 1 - valid
+    response = authorized_client.post(url, data)
+    assert response.status_code == 200
+    assert isinstance(response.context["form"], CaseAssignmentQueueForm)
+    assert response.context["wizard"]["steps"].current == "SELECT_QUEUE"
+
+    html = BeautifulSoup(response.content, "html.parser")
+    assert "Select a team queue to add the case to" in html.find("h1").get_text()
+
+    data = {
+        "case_assignments_case_assignee-current_step": "SELECT_QUEUE",
+    }
+    # POST step 2
+    response = authorized_client.post(url, data)
+    assert response.status_code == 200
+    assert response.context["form"]["queue"].errors == ["Select a queue to add the case to"]
+
+
+@pytest.fixture
+def mock_team_queue(requests_mock, data_queue):
+    data_queue["is_system_queue"] = False
+    url = client._build_absolute_uri("/queues/")
+    yield requests_mock.get(url=re.compile(f"{url}.*/"), json=data_queue)
+
+
+@pytest.fixture
+def mock_put_assignments(requests_mock, data_queue, data_assignment):
+    url = client._build_absolute_uri(f"/queues/{data_queue['id']}/case-assignments/")
+    return requests_mock.put(url=url, json=data_assignment)
+
+
+def test_case_assignments_add_user_team_queue(
+    authorized_client,
+    data_queue,
+    data_standard_case,
+    data_assignment,
+    mock_standard_case_with_assignments,
+    mock_team_queue,
+    mock_gov_users,
+):
+
+    case = data_standard_case
+    url_base = reverse("queues:case_assignments_assign_user", kwargs={"pk": data_queue["id"]})
+    url = f"{url_base}?cases={case['case']['id']}"
+    response = authorized_client.get(url)
+    assert response.status_code == 200
+    context = response.context
+    assert isinstance(context["form"], CaseAssignmentUsersForm)
+    # Flow initiated from a team queue, so expect no followup step
+    assert context["wizard"]["steps"].next == None
+
+    html = BeautifulSoup(response.content, "html.parser")
+    assert "Which users do you want to assign to this case?" in html.find("h1").get_text()
+
+
+def test_case_assignments_add_user_team_queue_submit_success(
+    authorized_client,
+    data_queue,
+    data_standard_case,
+    data_assignment,
+    mock_team_queue,
+    mock_gov_users,
+    mock_put_assignments,
+    mock_standard_case_with_assignments,
+    mock_standard_case_ecju_queries,
+    mock_standard_case_assigned_queues,
+    mock_standard_case_documents,
+    mock_standard_case_additional_contacts,
+    mock_standard_case_activity_filters,
+):
+
+    user_id = "1f288b81-2c26-439f-ac32-2a43c8b1a5cb"
+    case = data_standard_case
+    url_base = reverse("queues:case_assignments_assign_user", kwargs={"pk": data_queue["id"]})
+    url = f"{url_base}?cases={case['case']['id']}"
+    data = {
+        "case_assignments_case_assignee-current_step": "SELECT_USERS",
+        "SELECT_USERS-users": [user_id],
+        "SELECT_USERS-note": "foobar",
+    }
+    response = authorized_client.post(url, data, follow=True)
+    assert response.status_code == 200
+    assert response.redirect_chain[-1][0] == f"/queues/{data_queue['id']}/"
+    messages = [str(msg) for msg in response.context["messages"]]
+    expected_message = "Case adviser was added successfully"
+    assert messages == [expected_message]
+    assert mock_put_assignments.called
+    assert mock_put_assignments.last_request.json() == {
+        "case_assignments": [
+            {
+                "case_id": data_standard_case["case"]["id"],
+                "users": [user_id],
+            }
+        ],
+        "note": "foobar",
+        "remove_existing_assignments": False,
+    }
+
+    html = BeautifulSoup(response.content, "html.parser")
+    assert expected_message in html.select("div.app-snackbar__content")[0].get_text()
+
+
+def test_case_assignments_add_user_team_queue_submit_validation_error(
+    authorized_client,
+    data_queue,
+    data_standard_case,
+    data_assignment,
+    mock_team_queue,
+    mock_gov_users,
+    mock_put_assignments,
+    mock_standard_case_with_assignments,
+    mock_standard_case_ecju_queries,
+    mock_standard_case_assigned_queues,
+    mock_standard_case_documents,
+    mock_standard_case_additional_contacts,
+    mock_standard_case_activity_filters,
+):
+
+    user_id = "1f288b81-2c26-439f-ac32-2a43c8b1a5cb"
+    case = data_standard_case
+    url_base = reverse("queues:case_assignments_assign_user", kwargs={"pk": data_queue["id"]})
+    url = f"{url_base}?cases={case['case']['id']}"
+    data = {
+        "case_assignments_case_assignee-current_step": "SELECT_USERS",
+        "SELECT_USERS-users": [],
+        "SELECT_USERS-note": "foobar",
+    }
+    response = authorized_client.post(url, data)
+    assert response.status_code == 200
+    assert response.context["form"]["users"].errors == ["Select a user to allocate"]
